extracted mod data to new mock/tetsts/benchmarking files
